``` fix(util_torch): 修复张量转换中的数据类型问题
在tensorimg2ndarray和ImgTensor2NdarrayShowable函数中添加.float()调用，
确保张量在转换为numpy数组之前具有正确的浮点数据类型，避免潜在的数据类型不匹配错误。
```

+++ py/utilitypack/util_torch.py
@@ -185,7 +185,7 @@ def load_state_dict_ignore_tensor_unmatched(
 
 
 def tensorimg2ndarray(m: torch.Tensor):
-    m = m.cpu().numpy()
+    m = m.float().cpu().numpy()
     if not len(m.shape) == 2:  # not single channeled
         m = np.moveaxis(m, -3, -1)
     return m
@@ -917,7 +917,7 @@ def tensor_contains_nan(x: torch.Tensor):
 
 
 def ImgTensor2NdarrayShowable(x: torch.Tensor):
-    return x.cpu().numpy().transpose(0, 2, 3, 1).squeeze()
+    return x.float().cpu().numpy().transpose(0, 2, 3, 1).squeeze()
 
 
 class PositionalEmbeddingSinusoidal(torch.nn.Module):
